requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
dataclasses==0.7
Django==2.1.7
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djongo==1.2.31
idna==2.8
jsonschema==3.0.1
mongoengine==0.19.1
PyJWT==1.7.1
pymongo==3.7.2
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1.2
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.1
urllib3==1.24.1
uWSGI==2.0.18
